requirements.txt
boto3==1.18.33
psycopg2-binary==2.9.1
